Move openDesktop function from resource-list to utils/desktop.ts
