+++ libs/common/src/lib/resources/resource-list/resource-list.component.ts
@@ -9,7 +9,7 @@ import {
 } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
 import { TranslateService } from '@ngx-translate/core';
-import { BehaviorSubject, catchError, forkJoin, from, mergeMap, Observable, of, Subject, take, timer } from 'rxjs';
+import { BehaviorSubject, catchError, forkJoin, from, mergeMap, Observable, of, Subject, take } from 'rxjs';
 import { debounceTime, delay, filter, map, switchMap, takeUntil, tap, toArray } from 'rxjs/operators';
 import { ActivatedRoute, Router } from '@angular/router';
 import {
@@ -32,7 +32,7 @@ import { SisModalService, SisToastService } from '@nuclia/sistema';
 import { SampleDatasetService } from '../sample-dataset/sample-dataset.service';
 import { LabelsService } from '../../label/labels.service';
 import { OptionModel, PopoverDirective, TRANSITION_DURATION } from '@guillotinaweb/pastanaga-angular';
-import { LOCAL_STORAGE, WINDOW } from '@ng-web-apis/common';
+import { LOCAL_STORAGE } from '@ng-web-apis/common';
 import { getClassificationsPayload } from '../edit-resource';
 import {
   ColoredLabel,
@@ -42,8 +42,8 @@ import {
   ResourceWithLabels,
 } from './resource-list.model';
 import { UploadService } from '../../upload/upload.service';
-import { getDesktopAppUrl, getDesktopPlatform, RELEASE_URL } from '../../utils';
 import { NavigationService } from '../../services';
+import { openDesktop } from '../../utils';
 
 const POPOVER_DISPLAYED = 'NUCLIA_STATUS_POPOVER_DISPLAYED';
 
@@ -57,9 +57,10 @@ export class ResourceListComponent implements OnInit, OnDestroy {
   @ViewChild('failedPopoverDirective') failedPopoverDirective?: PopoverDirective;
 
   private localStorage = inject(LOCAL_STORAGE);
-  private window = inject(WINDOW);
   private sampleDatasetService = inject(SampleDatasetService);
   hasSampleData = this.sampleDatasetService.hasSampleResources();
+
+  unsubscribeAll = new Subject<void>();
 
   data: ResourceWithLabels[] | undefined;
   page = 0;
@@ -67,8 +68,6 @@ export class ResourceListComponent implements OnInit, OnDestroy {
   pageSize = DEFAULT_PAGE_SIZE;
   sort: SortOption = DEFAULT_SORTING;
   isLoading = true;
-  unsubscribeAll = new Subject<void>();
-  refreshing = true;
   searchOptions: OptionModel[] = [];
 
   statusCount = this.uploadService.statusCount.pipe(
@@ -694,31 +693,6 @@ export class ResourceListComponent implements OnInit, OnDestroy {
   }
 
   useDesktop() {
-    let appInstalled = false;
-
-    const onBlur = () => {
-      // A blur event indicates that the application is installed
-      appInstalled = true;
-    };
-    this.window.addEventListener('blur', onBlur);
-
-    timer(400).subscribe(() => {
-      if (!appInstalled) {
-        this.downloadDesktop();
-      }
-      this.window.removeEventListener('blur', onBlur);
-    });
-    this.window.location.href = 'nuclia-desktop://';
-  }
-
-  downloadDesktop() {
-    const platform = getDesktopPlatform();
-    if (platform) {
-      getDesktopAppUrl(platform).subscribe((url) => {
-        this.window.open(url || RELEASE_URL);
-      });
-    } else {
-      this.window.open(RELEASE_URL);
-    }
+    openDesktop();
   }
 }
